moved svgs to a separate directory and moved ServiceCard component correctly.
